notification table amended, documented endpoints added, plant_health_attributes endpoints (except the delete endpoint) finished

+++ plantapi/blueprints/data_provider/plant_health_attribute_data_provider/__init__.py
@@ -28,12 +28,14 @@ def getPlantHealthAttributeDtoById(plant_health_attribute_id):
     return session.query(Plant_Health_Attribute).get(plant_health_attribute_id)
 
 
-def deletePlantHealthAttributeDtoById(plant_health_attribute_id):
+def deletePlantHealthAttributeDtoById(plant_id, plant_health_attribute_id):
     plant_health_attribute = getPlantHealthAttributeDtoById(plant_health_attribute_id)
-    session.delete(plant_health_attribute)
-    session.commit()
+    if plant_health_attribute.plant_id == plant_id:
+        session.delete(plant_health_attribute)
+        session.commit()
+        return {'Plant Health Attribute deleted'}
 
-    return {'Plant Health Attribute ' + plant_health_attribute.health_attribute_id + ' deleted'}
+    return {}
 
 
 def updatePlantHealthAttributeDtoById(plant_health_attribute_id, upper_required_value, lower_required_value,
